requirements.txt
peft==0.7.0
dacite==1.8.1
h5py==3.7.0
numpy==1.23.5
protobuf==3.20.3
PyYAML==6.0
tensorflow==2.8.0
tensorflow-datasets==4.9.2
timm==0.9.5
torchmeta==1.8.0
torch==2.0.1+cu118
transformers==4.30.2
torchmetrics==1.0.1
tqdm==4.65.0
wandb==0.15.8
opencv-python==4.8.0.76
rich==13.4.2